[refactor] 타일 및 보드 자료형 변경

+++ src/component/board.rs
@@ -8,7 +8,7 @@ use crate::entity::defines::{TILE_HEIGHT, TILE_WIDTH};
 #[derive(Component, Debug, Clone, PartialEq)]
 pub struct Board {
     pub tiles: Vec<Tile>,
-    pub score: u32,
+    pub score: u16,
 }
 
 impl Board {
@@ -26,9 +26,10 @@ impl Board {
         if self.tiles.len() >= (TILE_WIDTH * TILE_HEIGHT) as usize {
             return;
         }
+        // TODO: Implement a better way to spawn tiles
         loop {
-            let x = (random::<u32>() % TILE_WIDTH as u32) as i32;
-            let y = (random::<u32>() % TILE_HEIGHT as u32) as i32;
+            let x = (random::<u32>() % TILE_WIDTH as u32) as u8;
+            let y = (random::<u32>() % TILE_HEIGHT as u32) as u8;
 
             if self.get_tile(x, y).is_none() {
                 let score = if random::<u32>() % 10 == 0 {
@@ -42,7 +43,7 @@ impl Board {
         }
     }
 
-    fn get_tile(&self, x: i32, y: i32) -> Option<&Tile> {
+    fn get_tile(&self, x: u8, y: u8) -> Option<&Tile> {
         self.tiles.iter().find(|tile| tile.position_x == x && tile.position_y == y)
     }
 }
